+++ test/utils.py
@@ -4,9 +4,9 @@ from unittest.mock import patch
 from hera.main import main
 
 
-def execute_program_helper(program):
+def execute_program_helper(program, *, flags=[]):
     with patch("sys.stdin", StringIO(program)):
-        return main(["--no-color", "-"])
+        return main(flags + ["--no-color", "-"])
 
 
 def preprocess_program_helper(program):